refactor: rename ContactPage to AboutPage
